ESWE-1464: Add validation to job updates (#108)
Adds the following to the update job journey:
- Validation on the CYA page of the 'Offence Exclusion Details' and
'Primary/Secondary Job Source' fields. The validation runs on page load
and when attempting to submit the form (i.e. on the GET and the POST).
- Displays error summary at the top of the page, with links to the
invalid field(s).
    - Displays inline error message with the field.
- Validation on page load of all the pages in the update job journey.

Defines a new 'govuk-summary-list__key--invalid' scss component to apply
the red line next to items in a summary list. This isn't supported in
the default govuk summary list class, so we needed to create a custom
class.