ESWE-1379 Job duplication journey - first half (#119)
This adds the first half of the job duplication journey:
- Adds 'duplicate' button to the CYA page for a job
- Add new 'duplicate job' page, pre-populated with data from the source
job, allowing the user to make changes to the job details, or to cancel
and return to the source job CYA page.

The second half of the journey will be implemented separately (see
https://dsdmoj.atlassian.net/browse/ESWE-1383), i.e. adding an 'are you
sure' page when the user clicks 'submit', and then saving the new job to
the database.